Extract a mixin for users of OvsClient instances
Move the setup code needed to instantiate OvsClient's to a mixin that
will be shared between scenarios and contexts.
